requirements.txt
future==0.18.2
opencv-python==4.5.5.62
numpy==1.20.3
tensorflow==2.8.0
dataclasses==0.8
six==1.16.0
sentencepiece==0.1.96
seqeval==1.2.2
sacrebleu==2.0.0
scipy==1.7.1
scikit-learn==0.24.2
pyyaml==5.4.1
pillow==8.4.0
pycocotools==2.0.4
keras==2.8.0
matplotlib==3.4.3
nltk==3.6.5
setuptools==58.0.4
mock==4.0.3
pandas==1.3.4
scikit-image==0.18.3
h5py==3.2.1
tensorboard==2.8.0
contextlib2==0.6.0.post1
ipython==7.29.0
lvis==0.5.3
lxml==4.6.3
networkx==2.6.3
yaml==0.2.5
zlib==1.2.11